requirements.txt
Django==1.11.13
django-cors-headers==2.2.0
djangorestframework==3.8.2
MySQL-python==1.2.5
pkg-resources==0.0.0
pytz==2018.4
